+++ config/cloud.cfg.tmpl
@@ -72,6 +72,7 @@ cloud_config_modules:
  - emit_upstart
  - snap_config
 {% endif %}
+ - setup-azure-networking
  - ssh-import-id
  - locale
  - set-passwords
@@ -125,7 +126,6 @@ cloud_final_modules:
  - phone-home
  - final-message
  - power-state-change
- - dhcpagent
  
 # System and/or distro specific settings
 # (not accessible to handlers/transforms)
